requirements.txt
Django>=1.7
South==0.8.1
django-betty-cropper>=0.1.4
djangorestframework==2.4.3
django-elastimorphic>=0.1.0
django_polymorphic==0.6.1
django-json-field==0.5.5
djangorestframework-csv==1.3.3
python-dateutil==2.1
pytz==2012h
requests>=1.1.0
simplejson==3.3.0
six==1.6.1
firebase-token-generator==1.3.2
